cadetDjangoModel/pyproject.toml
[project]
name = "cadetdjangomodel"
version = "0.1.0"
description = ""
authors = [
    {name = "PradoAllan",email = "allanprado.it@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.1.4,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
